Complete encryption help module creation. No testing added.
